i2c: add option to disable passthru protection
I2C passthrough protection allows passthrough to be disable for specific
slave addresses on each I2C port on the device. This can be used to
prevent the host from accessing devices that only the SCU should have
access to.

The way this passthrough protection is currently implemented does not
play well with I2C muxes. While it would be straightforward to fix this,
there's currently not a compelling need to be able to restrict on a
per-address basis. Instead, add a configuration option to disable
this passthrough protection.

Note that it's still possible to disable passthrough on a per-port
basis, which should be sufficient for our needs.

Signed-off-by: Michael Auchter <michael.auchter@ni.com>