chip: stm32: separate handling of AF and STOPF
The I2C slave will signal a STOPF (indicating that it's detected a STOP)
on concurrent with the last RxNE interrupt it receives for a write to
the slave.

Previously, this code reset the "tx_pending" flag if it detected a STOP;
when the master then started a read cycle, the slave would respond with
a constant `0xec` since "tx_pending" was cleared.

STOPF isn't a failure, it's an expected part of the transaction, so
don't clear out the pending flags when we receive it.

Signed-off-by: Michael Auchter <michael.auchter@ni.com>