system: don't immediately respond to nop reboot
The reboot host command handler will in some cases immediately send a
response to the host prior to scheduling a reboot, as after the reboot
the context is lost and no response is possible.

However, if the host requests a reboot to the already running image,
system_run_image_copy() considers this to be a no-op and will
immediately return to the caller instead of rebooting into the image. In
this case, execution continues and the hostcmd is responded to by
common/host_command.c as well. The duplicate response results in the
write failing:

/sys/class/chromeos/cros_ec# grep '^Firmware copy' version | \
	sed 's/.*: //' | tr -d '\n' > reboot
[  107.767243] cros-ec-i2c 0-003c: bad packet checksum
[  107.823315] cros-ec-i2c 0-003c: Command xfer error (err:-74)
tr: write error: Bad message

Don't immediately respond to the command if it'll be a no-op in order to
avoid sending a duplicate response to the host.

BUG=None
BRANCH=None
TEST=Triggering a reboot from the host into the currently running image
     doesn't error

Signed-off-by: Michael Auchter <michael.auchter@ni.com>
Change-Id: I921043c87cc498c9262310aa921ed41464a5ee31