titanium: add pwm fan control
Add support for pwm based fan control. This comprises
of controlling the fan speed using PWM signals generated using
timers in PWM mode. The sensing of the actual fan speed
using the tachometer signals from the fan is performed
by the timers in input capture mode. A control algorithm
is also implemented to be able to maintain fan speed.
Users can set the fan speed in rpm or percentage
using the fanset command and the current status of the
fans can be determined using the faninfo command.

Signed-off-by: Virendra Kakade <virendra.kakade@ni.com>
Change-Id: I83a7531ad6d5d19967d1148f4c7e588d8dbb2012