board: titanium: add daughterboard power control
This adds code for sequencing the supply rails to the daughtercards
based upon a defined power sequence. A specific daughterboard can be
powered on or off via the SCU console, or from the host. Additional
commands are implemented on the SCU console to show the state of the
rails and the power sequence, as well as to modify the sequence (for
debugging).

Signed-off-by: Michael Auchter <michael.auchter@ni.com>