board: titanium: add eeprom support
Add support for reading all SCU-connected EEPROMs which use a TLV format
for storage.

All EEPROMs are read into SCU memory and validated at init. A console
command is provided to dump the contents of each eeprom, either as a
human-readable dump of the TLV contents (for tags the SCU knows about),
or as a raw byte dump.

Signed-off-by: Michael Auchter <michael.auchter@ni.com>