titanium: add power button shutdown support
Add support for triggering a linux shutdown on the AP using a
POWER_BUTTON MKBP event.
Also add support for forced shutdown on a long press of power button.
A forced shutdown comprises of signalling AP to initiate shutdown and
then powering down all the voltage rails if the power button is still
in pressed position after a FORCED_SHUTDOWN_DELAY.

Signed-off-by: Virendra Kakade <virendra.kakade@ni.com>