board: titanium: use timer 5 for watchdog debug
There are two separate hwtimer implementations for the STM32: one for
16-bit timers, and one for 32-bit timers. CONFIG_STM_HWTIMER32 is set on
Titanium as we're using a 32-bit timer for the main clock (TIM2). The
aux timer that's used for watchdog debugging is configured in the same
file, and was previous set to TIM11, which is a 16-bit timer. Switch to
TIM5, a 32-bit timer, for consistency.

Signed-off-by: Michael Auchter <michael.auchter@ni.com>