ioexpander: maintain state across sysjumps
Since the reset signal resets the IO expanders and then triggers a soft
reset on the SCU, we need to make sure that we reinitialize the IO
expanders even on a soft reboot. This will be fixed on Rev. C where the
same reset line will reset the SCU.

For now, work around it by only skipping initialization after a sysjump.

Signed-off-by: Michael Auchter <michael.auchter@ni.com>