titanium: fix pwr_db led and power state errors
Remove the dependency of pwr_db LED state initialization on ioex_init_default
by forcefully setting them to default values on system reset. However
restrict this to only the cases in which sys_jump was not performed as in
sys_jump case we need to preserve old state values. Similar changes are
made to power state management.

Signed-off-by: Virendra Kakade <virendra.kakade@ni.com>