power: zynqmp: fix power rail sequence
Reorder the sequence to have a higher likelihood of successfully
bringing up the power rails.

There are a few different issues at play here, but the gist of it is
that all of the switching power supplies are clocked by the CPLD. The
power supplies default to using an internal clock until it sees a high
level on the clock pin, and then it switches to use the clock.

The oscillator that the clocks are derived from is on the 3.3V_CLK rail,
and the CPLD bank that received that clock is on the 2.5V rail. The
clocks to the 1.8V and 3.3V supplies are supposed to be gated by the
CPLD until the clock is stable, but that's not happening for reasons.
Furthermore, there's no gating for the clocks to other power supplies.

Enabling the 2.5V rail without having the clock running can cause a
glitch on the clocks to the power supplies, which causes them to switch
to external clock mode, but we don't supply a clock. This leads to the
power rails becoming unstable.

Kumaran has suggested this sequence for now, but it will likely change
again, and we'll need to do something better on Rev. C.

Signed-off-by: Michael Auchter <michael.auchter@ni.com>