common: add host control gpio support
The intent of this is to add support for exposing certain GPIOs (which
may be local GPIOs, or GPIOs on an external IO expander) to the host
machine for it to access/control. This is intended to be used for GPIOs
where the host is the primary user.

This implements a separate set of host commands, distinct from the
existing host control GPIO commands, which are more efficient and allow
access to both local and ioexpander GPIOs. Additionally, these commands
only permit the host to access GPIOs which are explicitly enumerated in
the host_control_gpios list. This prevents the host from accessing GPIOs
that it shouldn't have access to, as well as provides a stable
identifier for the GPIO (rather than relying on the gpio.inc ordering to
remain fixed).

Signed-off-by: Michael Auchter <michael.auchter@ni.com>