common: pwrsup: add power supply helpers
This adds some common infrastructure for handling power supplies and
power supply sequences.

This allows a board to define the list of power supplies in a single
place, similar to how GPIOs are handled, and abstracts away the details,
such as whether the supply is controlled by a GPIO or IOEX, or whether
power good for the supply is indicated by a signal or a voltage level.

Signed-off-by: Michael Auchter <michael.auchter@ni.com>