power: zynqmp: update master_pg handling
MASTER_PG is the AND of all power-goods, as well as interrupts from the
IO expanders used for power control. The IO expanders will pull
MASTER_PG low whenever an input bit on the IO expander changes, so we
can't just look at MASTER_PG to determine if power is actually good.

Instead, we'll call into the pwrsup_check_supplies function to validate
that the supplies we care about are still good.

Signed-off-by: Michael Auchter <michael.auchter@ni.com>
Change-Id: Ie823acc36a2a5ca852049794162b0d4c7876b8e1