power: zynqmp: pull defaults from eeprom
Pull the default boot device from the EEPROM to allow this to be set on
a persistent basis. This will allow a user to easily control whether the
device is booting off of eMMC or SD.

Additionally, implement support for autobooting. When the autoboot flag
is set in the EEPROM, the device will automatically boot as soon as
power is applied.

Signed-off-by: Michael Auchter <michael.auchter@ni.com>
Change-Id: If129b01f6231ec9ba08798fc947c382121971efd