board: titanium: add x410 thermal support
Add thermal algorithm for x4xx devices whose goal is to maintain various
temperature zones on the device at/below specified temperature setpoints.

This is achieved by monitoring all available temperature sensors on the device
motherboard and any attached daughterboards and comparing them to the
setpoints to determine a required cooling. This is done for each zone.
These cooling requirements are then combined into a single device-wide
aggregate cooling requirement using a weighted-average or maximum of all
policy. This cooling requirement is used to set the fan speeds to
achieve cooling.

Add a new console command tempzones which can be used to view
status of all zones on the device and also provide the capability to
modify the setpoints for all zones and the cooling weight associated
with them.

Signed-off-by: Virendra Kakade <virendra.kakade@ni.com>