temp_sensor: tmp112: support for multiple sensors
This adds support for multiple sensors to the tmp112 driver. If multiple
tmp112 sensors are present, this can be represented by adding a
tmp112_sensors array describing the ports and addresses of the sensors
to the the board file and defining TMP112_COUNT to be the length of this array.

For compatibility, if TMP112_COUNT is not defined by the board, the
driver falls back to the prior fixed I2C port and address.

Signed-off-by: Michael Auchter <michael.auchter@ni.com>
Change-Id: I7a28708423439cddb7c7faba7908a0ea2fdd6371