board: titanium: fix fan bugs and improvements
Fixes a bug where fans would report to be running even when they were
not; example: stalled fans or if fan connector is disconnected.
This was done by keeping track of fan rotations and adding a fan
monitor loop which checks if the number of fan rotations is
not incrementing.
Now the fan status, as reported by faninfo command reports rpm as 0
and status as "stopped" in above scenarios.

Also increased the lowest allowed fan speed to 4000 from 3800.
This was necessitated by the fact that even at the lowest
PWM duty cycle (1 percent) the fan speed was above 3800.
Previously the fan status would not stabilize to "locked"
at the lowest rpm settings.

Minor formatting changes were also made to the fan code.

Signed-off-by: Virendra Kakade <virendra.kakade@ni.com>