board: titanium: add thermal PID algorithm support
Re-haul the thermal regulation algorithm on the titanium board
based on feedback from thermal VnV.

1. Remove dependency of a zone's active cooling requirement on
t_warn setpoint. The main motivation of doing this is to
de-couple the effect of t_warn on the cooling calculation
and keep it only as a trigger for safety mechanism. This
provides more freedom in determining the correct setpoints
and weights for the thermal algorithm in order to enable
zones to achieve the t_target limits.

Old formula for zone's cooling reqd:
cool_req = 100 * (t_zone - t_target) / (t_warn - t_target)

New formula (PID algorithm):

cool_reqd = (t_zone - t_target) * kp + integral * ki

Where kp is the zone's proportional factor to multiply the "error" signal
(t_zone - t_target), ki is the integral coefficient and
"integral" is the running sum of the error signal.

kp and ki are expressed as a percentage.
e.g. If kp = 200 the P factor (in PID based control) effectively
becomes 200 / 100 = 2.

2. PID Algorithm Errors and Mitigations:

a. Overshoot at zero error:
When the PID algorithm manages to cool the device enough that the error signal
reaches zero it is possible that the integral value is high and will cause the
cooling requirement to still be high when it should be minimal. To mitigate this,
we reset the integral to zero when the error reaches a very small value/threshold
(typically 1 to 2 deg C). In reality, the error signal (T_Zone – T_Target) does
not reach and remain at zero; it rather stabilizes at a constant residual error
value. This is acceptable since we can have some margin to play with the
temperature setpoints.

b. Integral wind-up:
Integral is the running sum of the error signal. So, it is natural that
it can reach a very high unusable value. We mitigate this by restricting
the maximum value that we allow an integral to reach. This helps us control
the maximum contribution that the integral component can have on the
cooling requirement calculation.

3. Add new command to change the pid error constants like
minerror, maxerror and max allowed integral value.
Also provide a conditional to print out instantaneous and
average error values as a debug measure.

4. Add error averaging with a variable history length to alleviate
situtations where a high KP value causes a large fluctuation in
fan speed even with a degree of temperature change on an
important zone (zone with more weight).
This will correct for one-off spikes in error signal as
the averaging witll smooth it out.

Also multiply the error signal with a scalar before using it for
averaging. This gives the average error more precision and hence
results in smoother fan control.

5. Add ability to control error history length from console
using the piderr command. This allows us to control the
number of readings used for generating the rolling average.

Signed-off-by: Virendra Kakade <virendra.kakade@ni.com>