board: titanium: fixed cooling eeprom support
Add ability to cool the board at fixed cooling capacity
if so specified in the eeprom.
If the setting is present in the eeprom and in the valid range
(0-100) use it, otherwise default to the thermal algorithm.
When fixed cooling is in use, the thermal algorithm is disabled
and only the critical protection remains active.

Signed-off-by: Virendra Kakade <virendra.kakade@ni.com>