board: titanium: increase watchdog timeout
On some titanium boards we observed failures in the swsync process in
UBoot leading to failure. This was root-caused to be some units
taking longer than expected time to erase the RW portion of
flash (sectors 6 and 7).
Having a larger timeout value helps mitigate this situation and the
affected units did not show the same behavior again.

Signed-off-by: Virendra Kakade <virendra.kakade@ni.com>