neon: power: fix watchdog when booting the first time
With a190641b34, a change was introduced which unfortunately
does not work if the device booted for the first time after
power cycle. In this case the, the firmware is changed from
RO to RW while remaining in S0 power state. In this case,
the wdt_enabled flag was not set. As a consequence, actual
watchdog events caused by system crashes were ignored.

Signed-off-by: Joerg Hofrichter <joerg.hofrichter@ni.com>